+++ VisualStudio/ABB.SrcML.VisualStudio.SrcMLService.IntegrationTests/SrcMLServiceCPlusPlusTests.cs
@@ -12,6 +12,8 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 
 namespace ABB.SrcML.VisualStudio.SrcMLService.IntegrationTests {
+    using File = Pri.LongPath.File;
+    using Path = Pri.LongPath.Path;
 
     [TestClass]
     public class SrcMLServiceCPlusPlusTests : IInvoker {
@@ -150,7 +152,7 @@ namespace ABB.SrcML.VisualStudio.SrcMLService.IntegrationTests {
             TestSolution.Remove(addedProject);
 
             foreach(var expectedFile in expectedFiles) {
-                File.Delete(expectedFile);
+                if (File.Exists(expectedFile)) File.Delete(expectedFile);
                 Assert.IsTrue(resetEvent.WaitOne(scanIntervalMs * 2));
                 Assert.IsFalse(archive.ContainsFile(expectedFile));
             }
